refactor codec tests
