+++ tests/v3/test_codecs/test_gzip.py
@@ -1,0 +1,24 @@
+import numpy as np
+import pytest
+
+from zarr.abc.store import Store
+from zarr.array import Array
+from zarr.codecs import BytesCodec, GzipCodec
+from zarr.store.core import StorePath
+
+
+@pytest.mark.parametrize("store", ("local", "memory"), indirect=["store"])
+def test_gzip(store: Store) -> None:
+    data = np.arange(0, 256, dtype="uint16").reshape((16, 16))
+
+    a = Array.create(
+        StorePath(store),
+        shape=data.shape,
+        chunk_shape=(16, 16),
+        dtype=data.dtype,
+        fill_value=0,
+        codecs=[BytesCodec(), GzipCodec()],
+    )
+
+    a[:, :] = data
+    assert np.array_equal(data, a[:, :])
